+++ local-store/ClearFullStore.ts
@@ -35,7 +35,7 @@ class ClearFullStore {
         if (logInfo) {
             var end = UniqueChronologicalKeys.getMillisecondTime();
         }
-        if (this.itemsRemovedCallback) {
+        if (this.itemsRemovedCallback && removedItems.removedCount > 0) {
             this.itemsRemovedCallback(store, removedItems.items, {
                 error: err,
                 message: "removed " + removedItems.items.length + " local store entries" + (start || end ? " in " + Math.round(end - start) + " ms" : "") + ", " +