+++ test/local-store/LocalStoreByCategoryTest.ts
@@ -81,7 +81,7 @@ QUnit.test("local-store-by-category-1", function LocalStoreByCategoryScenario1Te
 
     sr.equal(store.rootStore.hasItem("-"), false);
 
-    assertDataStores(sr, store, [
+    assertDataStores(sr, store, false, compareCategoryEvent, [
         { name: "alpha", size: 2, data: [valB, valC] },
         { name: "omega", size: 2, data: [valD, valE] },
     ]);
@@ -89,7 +89,7 @@ QUnit.test("local-store-by-category-1", function LocalStoreByCategoryScenario1Te
 
     store.stores.alpha.removeItem(keyA1);
 
-    assertDataStores(sr, store, [
+    assertDataStores(sr, store, false, compareCategoryEvent, [
         { name: "alpha", size: 1, data: [valC] },
         { name: "omega", size: 2, data: [valD, valE] },
     ]);
@@ -113,18 +113,34 @@ QUnit.test("local-store-by-category-full-1", function localStoreByCategoryFullTe
     store.stores.delta.addItem("too-long", true);
 
     sr.equal(cI, 1);
+    sr.equal(dI, 0);
+
+    assertDataStores(sr, store, true, (s1, s2) => s1 === s2, [
+        { name: "delta", size: 1, data: ["too-long"] },
+    ]);
+
+    store.stores.charlie.addItem("A", true);
+
+    sr.equal(cI, 1);
+    sr.equal(dI, 0);
+
+    // charlie: ["five-", "A"]
+    // delta: ["too-long"]
+
+    store.stores.delta.addItem("two", true);
+
+    sr.equal(cI, 2);
     sr.equal(dI, 1);
 
-    //assertDataStores(sr, store, [
-    //    { name: "charlie", size: 0, data: [] },
-    //    { name: "delta", size: 0, data: [] },
-    //]);
+    assertDataStores(sr, store, true, (s1, s2) => s1 === s2, [
+        { name: "delta", size: 1, data: ["two"] },
+    ]);
 });
 
 
 /** Check that the TestStore contains the expected stores with expected sizes and data
  */
-function assertDataStores(sr: QUnitAssert, store: LocalStoreByCategory<any>, expectedSubStores: { name: string; size: number; data: any[] }[]) {
+function assertDataStores(sr: QUnitAssert, store: LocalStoreByCategory<any>, plainStr: boolean, compare: (val1, val2) => boolean, expectedSubStores: { name: string; size: number; data: any[] }[]) {
     var stores = store.getStoresContainingData();
     var names: string[] = [];
 
@@ -138,8 +154,9 @@ function assertDataStores(sr: QUnitAssert, store: LocalStoreByCategory<any>, exp
 
         if (expected.data && expected.data.length > 0) {
             for (var k = 0, sizeK = expected.data.length; k < sizeK; k++) {
-                sr.equal(CommonStorageTests.looseEqual(compareCategoryEvent, categoryStore.getData(), expected.data), true,
-                    "expected: " + JSON.stringify(expected.data) + ", actual: " + JSON.stringify(categoryStore.getData()));
+                var storeData = categoryStore.getData(plainStr);
+                sr.equal(CommonStorageTests.looseEqual(compare, storeData, expected.data), true,
+                    "expected: " + JSON.stringify(expected.data) + ", actual: " + JSON.stringify(storeData));
             }
         }
     }